atualizar

+++ ODSU/ODSU/src/main/java/com/joaquimFelix/ODSU/service/oficina/OficinaServiceImpl.java
@@ -3,6 +3,7 @@ package com.joaquimFelix.ODSU.service.oficina;
 
 import com.joaquimFelix.ODSU.model.entity.Oficina;
 import com.joaquimFelix.ODSU.repository.oficina.OficinaRepository;
+import jakarta.persistence.EntityNotFoundException;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
@@ -21,12 +22,18 @@ public class OficinaServiceImpl implements OficinaService {
 
     @Override
     public List<Oficina> listarOficina() {
+        try {
         return oficinaRepository.findAll();
+        } catch (Exception e) {
+            // Trate a exceção aqui
+            throw new RuntimeException("Erro ao listar oficinas", e);
+        }
     }
 
     @Override
     public Oficina obterOficinaPorId(Long id) {
-        return oficinaRepository.findById(id).orElse(null);
+        return oficinaRepository.findById(id)
+                .orElseThrow(() -> new EntityNotFoundException("Oficina não encontrada com o ID: " + id));
     }
 
     @Override
@@ -48,12 +55,17 @@ public class OficinaServiceImpl implements OficinaService {
             oficinaAtualizada.setCnpj(oficina.getCnpj());
             return oficinaRepository.save(oficinaAtualizada);
         } else {
-            return null;
+            throw new EntityNotFoundException("Oficina não encontrada");
         }
     }
 
     @Override
     public void excluirOficina(Long id) {
+        // Verifica se a oficina existe antes de tentar excluí-la
+        if (!oficinaRepository.existsById(id)) {
+            throw new EntityNotFoundException("Oficina não encontrada");
+        }
+        // Se a oficina existe, então exclui
         oficinaRepository.deleteById(id);
     }
 }
